requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
pyaudio==0.2.14
librosa==0.11.0
speechrecognition==3.14.3
pvporcupine==3.0.5
pvcobra==2.0.5
pvrecorder==1.2.7
pvspeaker==1.0.5
soundfile==0.13.1
openai==1.99.9
python-dotenv
faster-whisper==1.2.0
numpy==1.26.4
colorlog
onnxruntime==1.22.1
onnxruntime-gpu==1.22.0
sounddevice>=0.4.4
snac>=1.2.1
coqui-tts==0.27.0
kokoro==0.9.4
ddgs==9.5.2
fandom-py==0.2.1
wikipedia==1.4.0
playwright==1.54.0
beautifulsoup4
langchain==0.3.27
langchain-community==0.3.27
sentence-transformers==5.1.0
chromadb==1.0.16
pandas==2.3.1
huggingface-hub==0.34.4
transformers==4.55.0 
triton-windows==3.2.0.post13